refactor: Response  StatusLine 이름 변경 및 패키지 구조 변경
