+++ tomcat/src/main/java/org/apache/coyote/controller/RegisterController.java
@@ -4,9 +4,9 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.domain.request.HttpRequest;
 import org.apache.coyote.domain.request.RequestBody;
-import org.apache.coyote.domain.response.ContentType;
+import org.apache.coyote.domain.response.statusline.ContentType;
 import org.apache.coyote.domain.response.HttpResponse;
-import org.apache.coyote.domain.response.HttpStatusCode;
+import org.apache.coyote.domain.response.statusline.HttpStatusCode;
 import org.apache.coyote.domain.response.RedirectUrl;
 import org.apache.coyote.domain.response.ResponseBody;
 import org.slf4j.Logger;
