+++ tomcat/src/main/java/org/apache/coyote/domain/response/HttpResponse.java
@@ -1,15 +1,14 @@
 package org.apache.coyote.domain.response;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
-import org.apache.coyote.domain.HttpCookie;
 import org.apache.coyote.domain.request.requestline.HttpVersion;
+import org.apache.coyote.domain.response.statusline.HttpStatusCode;
+import org.apache.coyote.domain.response.statusline.StatusLine;
 
 public class HttpResponse {
 
-    private ResponseLine responseLine;
+    private StatusLine statusLine;
     private ResponseBody responseBody;
     private final List<String> responseHeader;
 
@@ -17,15 +16,15 @@ public class HttpResponse {
         this.responseHeader = new ArrayList<>();
     }
 
-    public HttpResponse(final ResponseLine responseLine,
+    public HttpResponse(final StatusLine statusLine,
                         final ResponseBody responseBody) {
-        this.responseLine = responseLine;
+        this.statusLine = statusLine;
         this.responseBody = responseBody;
         this.responseHeader = new ArrayList<>();
     }
 
     public HttpResponse responseLine(HttpVersion httpVersion, HttpStatusCode httpStatusCode) {
-        this.responseLine = ResponseLine.of(httpVersion, httpStatusCode);
+        this.statusLine = StatusLine.of(httpVersion, httpStatusCode);
         return this;
     }
 
@@ -41,9 +40,7 @@ public class HttpResponse {
 
     public String getValue() {
         StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append("\r\n")
-                .append(responseLine.generateResponseString())
-                .append("\r\n");
+        stringBuilder.append(statusLine.generateResponseString());
         for (String header : responseHeader) {
             stringBuilder.append(header);
         }
